+++ src/frontend/src/views/index.ts
@@ -5,6 +5,7 @@ export * from "./HomeView";
 export * from "./LlmPromptView";
 export * from "./AgentCanvasView";
 export * from "./AgentMarketplaceView";
+export * from "./UsageView";
 export * from "./LandingPage";
 export * from "./LoginView";
 export * from "./RegisterView";